feat: Add half-day holidays feature for December 24 & 31
Implements German employment benefit where Christmas Eve and New Year's Eve
count as 0.5 vacation days instead of 1.0 when taken off.

Backend Changes:
- Add halfDayHolidaysEnabled field to User entity with database migration V13
- Modify WorkingDaysCalculator to return BigDecimal and support fractional days
- Add isHalfDayHoliday() helper to detect Dec 24/31
- Fix recurring off-days exclusion logic for vacation entries
  * Previously skipped recurring off-day checks for ALL time-off entries
  * Now only skips for sick/personal days, always checks for vacation
  * Ensures consistent day counts between balance card and time-off list
- Update TimeOffMapper to use .doubleValue() instead of .intValue()
- Change OpenAPI TimeOffResponse.days from integer to number (double)
- Add comprehensive test suite: 11 unit tests + 8 integration tests

Frontend Changes:
- Add checkbox with tooltip in ProfileView and AdminUsersView
- Display fractional days with 1 decimal place (e.g., "3.5 days")
- Update TimeOffView to show .toFixed(1) for all day counts
- Add translated "half-day" indicator in calendar tooltips
- Fix calendar colors (was broken due to undefined variable)
- Add i18n: "halber Tag" (DE) / "half-day" (EN)

All 162 tests passing.

+++ backend/src/main/java/cc/remer/timetrack/usecase/timeoff/TimeOffMapper.java
@@ -33,15 +33,16 @@ public class TimeOffMapper {
         response.setTimeOffType(TimeOffResponse.TimeOffTypeEnum.fromValue(entity.getTimeOffType().name()));
 
         // Calculate working days (excludes weekends, holidays, recurring off-days)
+        // Supports half-day holidays (Dec 24 & 31) which count as 0.5 days
         GermanState userState = entity.getUser().getState();
-        int workingDays = workingDaysCalculator.calculateWorkingDays(
+        BigDecimal workingDays = workingDaysCalculator.calculateWorkingDays(
                 entity.getUser().getId(),
                 userState,
                 entity.getStartDate(),
                 entity.getEndDate(),
                 entity.getId() // Exclude this entry from the calculation
         );
-        response.setDays(workingDays);
+        response.setDays(workingDays.doubleValue());
 
         response.setHoursPerDay(entity.getHoursPerDay() != null ? entity.getHoursPerDay().doubleValue() : null);
         response.setNotes(entity.getNotes());
